Refactor: move raft message types to separate files
- `raft::message::AppendEntriesRequest`
- `raft::message::AppendEntriesResponse`
- `raft::message::ClientWriteResponse`
- `raft::message::InstallSnapshotRequest`
- `raft::message::InstallSnapshotResponse`
- `raft::message::VoteRequest`
- `raft::message::VoteResponse`
