+++ openraft/src/raft/message/client_write.rs
@@ -1,0 +1,42 @@
+use std::fmt::Debug;
+
+use crate::AppDataResponse;
+use crate::LogId;
+use crate::Membership;
+use crate::MessageSummary;
+use crate::RaftTypeConfig;
+
+/// The response to a client-request.
+#[cfg_attr(
+    feature = "serde",
+    derive(serde::Deserialize, serde::Serialize),
+    serde(bound = "C::R: AppDataResponse")
+)]
+pub struct ClientWriteResponse<C: RaftTypeConfig> {
+    /// The id of the log that is applied.
+    pub log_id: LogId<C::NodeId>,
+
+    /// Application specific response data.
+    pub data: C::R,
+
+    /// If the log entry is a change-membership entry.
+    pub membership: Option<Membership<C::NodeId, C::Node>>,
+}
+
+impl<C: RaftTypeConfig> Debug for ClientWriteResponse<C>
+where C::R: Debug
+{
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("ClientWriteResponse")
+            .field("log_id", &self.log_id)
+            .field("data", &self.data)
+            .field("membership", &self.membership)
+            .finish()
+    }
+}
+
+impl<C: RaftTypeConfig> MessageSummary<ClientWriteResponse<C>> for ClientWriteResponse<C> {
+    fn summary(&self) -> String {
+        format!("log_id: {}, membership: {:?}", self.log_id, self.membership)
+    }
+}
